KnowAgent/pyproject.toml
[project]
name = "knowagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-ollama>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "tiktoken>=0.11.0",
    "wikipedia>=1.4.0",
]
